fix: mlx_utilを切り出し
