+++ Makefile
@@ -11,7 +11,7 @@ XLIBFT_SRCS = $(addprefix srcs/xlibft/, \
 		ft_xcalloc.c ft_xlstnew.c ft_xsplit.c ft_xstrdup.c ft_xstrjoin.c \
 		)
 UTILS_SRCS = $(addprefix srcs/utils/, \
-		error.c util.c \
+		error.c mlx_util.c util.c \
 		)
 SRCS = main.c $(VECTOR3_SRCS) $(XSYSCALL_SRCS) $(XLIBFT_SRCS) $(UTILS_SRCS)
 OBJS = $(SRCS:%.c=%.o)
