+++ includes/xlibft.h
@@ -6,7 +6,7 @@
 /*   By: takkatao <takkatao@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/01/20 23:39:45 by ahayashi          #+#    #+#             */
-/*   Updated: 2022/02/25 14:08:33 by takkatao         ###   ########.fr       */
+/*   Updated: 2022/05/16 11:08:19 by ahayashi         ###   ########.jp       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 # define XLIBFT_H
 
 # include "libft.h"
-# include "define.h"
+# include <stdio.h>
 
 char	*ft_xcalloc(size_t count, size_t size);
 t_list	*ft_xlstnew(void *content);
