+++ main.c
@@ -12,16 +12,6 @@
 
 #include "minirt.h"
 
-void	pixel_put_to_image(t_image *data, int x, int y, int color)
-{
-	char	*dst;
-	int		offset;
-
-	offset = (y * data->bytes_per_line + x * (data->bits_per_pixel / 8));
-	dst = data->data_addr + offset;
-	*(unsigned int *)dst = color;
-}
-
 /*
  * スクリーン座標をワールド座標に変換。
  * minilibXのpixelを3次元上のスクリーンに変換。
