+++ srcs/utils/mlx_util.c
@@ -1,25 +1,23 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   utils.h                                            :+:      :+:    :+:   */
+/*   mlx_util.c                                         :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
 /*   By: ahayashi <ahayashi@student.42tokyo.jp>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/03/04 17:02:57 by ahayashi          #+#    #+#             */
-/*   Updated: 2022/05/16 10:57:10 by ahayashi         ###   ########.jp       */
+/*   Created: 2022/05/16 10:54:43 by ahayashi          #+#    #+#             */
+/*   Updated: 2022/05/16 10:54:43 by ahayashi         ###   ########.jp       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#ifndef UTILS_H
-# define UTILS_H
+#include "utils.h"
 
-# include "define.h"
-# include "libft.h"
-# include "xlibft.h"
+void	pixel_put_to_image(t_image *data, int x, int y, int color)
+{
+	char	*dst;
+	int		offset;
 
-void	puterr(char *target, char *message);
-void	puterr_exit(char *target, char *message, int exit_status);
-char	*triple_join(char *s1, char *s2, char *s3);
-void	pixel_put_to_image(t_image *data, int x, int y, int color);
-
-#endif
+	offset = (y * data->bytes_per_line + x * (data->bits_per_pixel / 8));
+	dst = data->data_addr + offset;
+	*(unsigned int *)dst = color;
+}
